stage3: restructure into a folder-based layout
It's an absolute mess, but should make integrating stage3_revamp easier…
