+++ src/stages/stage3/meson.build
@@ -1,0 +1,15 @@
+stage3_src = files(
+    'background_anim.c',
+    'wriggle.c',
+    'scuttle.c',
+    'draw.c',
+    'nonspells/boss_nonspells.c',
+    'nonspells/midboss_nonspell_1.c',
+    'spells/deadly_dance.c',
+    'spells/firefly_storm.c',
+    'spells/light_singularity.c',
+    'spells/moonlight_rocket.c',
+    'spells/night_ignite.c',
+    'stage3.c',
+    'timeline.c',
+)
